rewrite porzucone.py for efficiency and memory use
It used to be very inefficient, as all the pages were collected
in a list and written at the end of script execution, which used
a lot of memory and was vulnerable to crashing. #11 says I should
rewrite this script using the dump data, but this should be a good
enough solution — I'll be monitoring it for a couple of weeks and
if it doesn't work properly, then I'll think about rewriting again.