deconvoluted morfAnalyse, imporved wikilink()
Maybe I lost something in the process (in morfAnalyse), but it doesn't
seem so for now. It just shows how lame my programming was 5 years
ago, and probably how little it has improved since then. I left the
"check in sjp.pl" part commented, just to have a look at it in future.