changed definition format, disabled next for last elem
I don't think there is a point in having definitions in json
as a list, especially that it makes it harder when there are multiple
parts of speech. So now it is just one chunk of text. Numbers (for the
selector) are present as another field, 'def_nums'.

Also disabled the "next" button if there's only one word on page.