check for edit timestamps
at data generation, record timestamp in "fetch_time" field. When the
bot verifies examples, it will check if there have been no edits to that
page (i.e. the one that we are adding an example to) after data
fetching. Ideally we should check if the examples field has not changed,
but this will do for now