method for joining sentences from utt_tagged
In "/select", utt_tagged result is an xml-tagged string. In order to
join a sentence and remove superfluous spaces ("przyjeżdżała" + "m") I'm
parsing it witih xmltodict first and then looping through to see if
there are any agglutinates. It is essentialy the same thing as I had in
this funciton earlier, but for a different input format.